+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+## Version 0.1.21
+> - Added hasSubStore method in substore.
+> - Subscribers get their own copy of the state.
+
 ## Version 0.1.20
 > - Added Substores
 > - Fixed Rollback.